move dataloader getters to base class
